Added code to keep larger numbers of objects in `meta` as well as photometry tables along with a selection threshold.
Now, we have the following possibilities:
- `keep_all_simulated_meta` will keep the metadata of rejected and not
observed SN if True, and only detected ones if False.
- With defaults it will assign selected column of -1 if not observed, 0
if not detected by simsurvey,  2 if passes the metadata selection criteria.
- By default light curves are only kept if the SN is selected by `meta_selection_func`.
If this is set to `None` the user must change `threshold_for_lc` to 1
for non-empty light curves.
- Added tests to ensure the working of new functionality. These tests
test the new cases.
- Added xfail for version of `simsurvey`.
This was necessary as the code uses a feature of simsurvey that is not in the current pip
release at version 0.6.0. It is implemented in the next release.
- Bump version

	modified:   tdd/io_simsurvey.py
	modified:   tests/test_simsurvey.py
	modified:   tdd/version.py